chore(test): 테스트코드 작성 및 폴더구조 수정
